resolving merge conflict, removed unused Operating system conditional, some refactoring to make the AutoConfiguration class simple
