--- azure-application-insights-spring-boot-starter/src/main/java/com/microsoft/applicationinsights/boot/conditional/OnOperationSystemCondition.java
@@ -1,55 +1,0 @@
-/*
- * ApplicationInsights-Java
- * Copyright (c) Microsoft Corporation
- * All rights reserved.
- *
- * MIT License
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this
- * software and associated documentation files (the ""Software""), to deal in the Software
- * without restriction, including without limitation the rights to use, copy, modify, merge,
- * publish, distribute, sublicense, and/or sell copies of the Software, and to permit
- * persons to whom the Software is furnished to do so, subject to the following conditions:
- * The above copyright notice and this permission notice shall be included in all copies or
- * substantial portions of the Software.
- * THE SOFTWARE IS PROVIDED *AS IS*, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
- * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR
- * PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE
- * FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
- * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
- * DEALINGS IN THE SOFTWARE.
- */
-
-package com.microsoft.applicationinsights.boot.conditional;
-
-import com.microsoft.applicationinsights.internal.system.SystemInformation;
-import org.springframework.boot.autoconfigure.condition.ConditionMessage;
-import org.springframework.boot.autoconfigure.condition.ConditionOutcome;
-import org.springframework.boot.autoconfigure.condition.SpringBootCondition;
-import org.springframework.context.annotation.Condition;
-import org.springframework.context.annotation.ConditionContext;
-import org.springframework.core.type.AnnotatedTypeMetadata;
-
-import java.util.Map;
-
-/**
- * {@link Condition} that checks if the application is running on specific operating system.
- *
- * @author Arthur Gavlyukovskiy
- */
-public class OnOperationSystemCondition extends SpringBootCondition {
-
-    @Override
-    public ConditionOutcome getMatchOutcome(ConditionContext context, AnnotatedTypeMetadata metadata) {
-        Map<String, Object> attributes = metadata.getAnnotationAttributes(ConditionalOnOperatingSystem.class.getName());
-        OperatingSystem operatingSystem = (OperatingSystem) attributes.get("value");
-        ConditionMessage.Builder message = ConditionMessage.forCondition(ConditionalOnOperatingSystem.class);
-        String name = operatingSystem.name();
-        if (operatingSystem == OperatingSystem.WINDOWS && SystemInformation.INSTANCE.isWindows()) {
-            return ConditionOutcome.match(message.foundExactly(name));
-        }
-        if (operatingSystem == OperatingSystem.UNIX && SystemInformation.INSTANCE.isUnix()) {
-            return ConditionOutcome.match(message.foundExactly(name));
-        }
-        return ConditionOutcome.noMatch(message.didNotFind(name).atAll());
-    }
-}
